# API_documentation.yaml
openapi: "3.0.0"
info:
  title: Review Guru API
  description: Back end API for Review Guru
  version: 0.0.1

servers:
  - url: https://www.api.harusbisa.net
    description: Production server
  - url: https://www.api.harusbisa.net/dev
    description: Development server
  
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    LoginResponse:
      properties:
        token:
          type: string
          example: '300b524e1b87374b56af68cd6b54aef7a65fda38'
        userId:
          type: string
          example: '5db6093a7f75df3ce6c12095'
            
    User:
      properties:
        userId:
          type: string
          example: '5e01c0379a3aba19b056f844'
        name:
          type: string
          example: 'John Doe'
        email:
          type: string
          example: 'john.doe@gmail.com'
        isVerified:
          type: boolean
          example: true
        outgoingReviews:
          type: array
          items:
            type: object
            properties:
              revieweeId:
                type: string
                example: '5e01c0379a3aba19b056f844'
              reviewId:
                type: string
                example: '5e17b09bcf138f14b804c386'
        helpfulnessVotes:
          type: array
          items:
            type: object
            properties:
              revieweeId:
                type: string
                example: '5e01c0379a3aba19b056f844'
              reviewId:
                type: string
                example: '5e17b09bcf138f14b804c386'
              vote:
                type: string
                example: 'upVote'
          
    Review:
      properties:
        reviewId:
          type: string
          example: '5db492c349a67b33b8d0a2a2'
        courseName:
          type: string
          example: 'Brand Management 101'
        overallRating:
          type: number
          example: 4.5
        recommendationRating:
          type: number
          example: 2.5
        difficultyRating:
          type: number
          example: 3.5
        textbookRequired:
          type: boolean
          example: true
        yearTaken:
          type: integer
          example: 2015
        grade:
          type: string
          example: 'A-'
        review:
          type: string
          example: 'This class is so easy and the professor teaching style makes me understand the course material easily.'
        helpfulUpVote:
          type: integer
          example: 25
        helpfulDownVote:
          type: integer
          example: 2
        userVote:
          type: string
          example: 'upVote'
        isAuthor:
          type: boolean
          example: true
        tags:
          type: array
          items:
            type: string
          example: ['caring', 'young', 'easy class']
        teachingStyles:
          type: array
          items:
            type: string
          example: ['visual', 'professional']
      
    Reviewee:
      allOf:
        - $ref: '#/components/schemas/RevieweeMatch'
        - type: object
          properties:
            overallRating:
              type: number
              example: 4.5
            recommendationRating:
              type: number
              example: 5
            difficultyRating:
              type: number
              example: 4.5
            numberOfReviews:
              type: string
              example: 12
            reviews:
              type: array
              items:
                $ref: '#/components/schemas/Review'

    RevieweeMatch:
      properties:
        revieweeId:
          type: string
          example: '5db492c349a67b33b8d0a2a2'
        name:
          type: string
          example: 'Timothy Bryan'
        school:
          type: string
          example: 'Universitas Prasetya Mulya'

paths:
  /signup:
    post:
      summary: Create a new user.
      tags: 
      - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
             type: object
             required:
              - name
              - email
              - password
             properties:
              name:
                type: string
                example: 'Wilson Burnawan'
              email:
                type: string
                format: email
                example: 'wilson.burnawan@uiuc.edu'
              password:
                type: string
                example: '123abc'
      responses:
        '201':
          description: Create user is successful.
        '500':
          description: There was an error creating the user.
  
  /login:
    post:
      summary: Login to get authorization token.
      tags:
      - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: 
              - email
              - password
              properties:
                email:
                  type: string
                  format: email
                  example: 'wilson.burnawan@uiuc.edu'
                password:
                  type: string
                  example: '123abc'
      responses:
        '200':
          description: Login is successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Please provide correct email and password.
        '500':
          description: There was an error logging in.
  
  /verification/{token}:
    get:
      summary: Verify user account.
      tags:
       - Authentication
      parameters: 
        - in: path
          required: true
          name: token
          description: Temporary token given to the user to verify the account.
          schema:
            type: string
            example: '5db492c349a67b33b8d0a2a2'
      responses:
        '301':
          description: Redirect the user to the login page.
        '404':
          description: We were unable to find a user for this token.
  
  /resend:
    post:
      summary: Resend verification email to the user.
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
             type: object
             required:
              - email
             properties:
              email:
                type: string
                example: 'john.doe@email.com'
      responses:
        '301':
          description: Redirect the user to the login page.
        '404':
          description: We were unable to find a user with the given email.
          
  /users/{userId}:
    get:
      summary: Get user data by id.
      tags:
      - User
      security: 
        - BearerAuth: []
      parameters: 
        - in: path
          required: true
          name: userId
          description: Id of the user.
          schema:
            type: string
            example: '5db492c349a67b33b8d0a2a2'
      responses:
        '200':
          description: Get user is successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Please authenticate with the correct credential.
        '404':
          description: User not found.
        '500':
          description: There was an error getting the user.
      
  /reviewees:
    post:
      summary: Create a new review for new reviewee.
      tags: 
      - General
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
             type: object
             required:
              - name
              - school
              - courseName
              - overallRating
              - recommendationRating
              - difficultyRating
              - grade
              - yearTaken
              - review
              - textbookRequired

             properties:
              name:
                type: string
                example: 'Wilson Burnawan'
              school:
                type: string
                example: 'Universitas Prasetya Mulya'
              courseName:
                type: string
                example: 'Brand Management 101'
              overallRating:
                type: number
                example: 4.5
              recommendationRating:
                type: number
                example: 2.5
              difficultyRating:
                type: number
                example: 3.5
              textbookRequired:
                type: boolean
                example: true
              yearTaken:
                type: integer
                example: 2015
              grade:
                type: string
                example: 'A-'
              review:
                type: string
                example: 'This class is so easy and the professor teaching style makes me understand the course material easily.'
              tags:
                type: array
                items:
                  type: string
                example: ['caring', 'young', 'easy class']
              teachingStyles:
                type: array
                items:
                  type: string
                example: ['visual', 'professional']       
      responses:
        '201':
          description: Create review for new reviewee is successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reviewee'
        '401':
          description: Authentication is required to create new review.
        '500':
          description: There was an error creating a review for new reviewee.
    
    get:
      summary: Get list of a reviewees matching the query.
      tags:
        - General
      parameters: 
        - in: query
          name: name
          description: Name of the reviewee.
          schema:
            type: string
            example: 'Timothy Bryan'
      responses:
        '200':
          description: Get list of reviewees matching the query is successful.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RevieweeMatch'
        '500':
          description: There was an error getting the reviewees matching the query.
   
  /reviewees/{revieweeId}:
    get:
      summary: Get a specific reviewee.
      tags:
        - Specific Reviewee
      parameters: 
        - in: path
          required: true
          name: revieweeId
          description: Id of reviewee.
          schema:
            type: string
            example: '5db492c349a67b33b8d0a2a2'
      responses:
        '200':
          description: Get reviewee is successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reviewee'
        '404':
          description: Reviewee not found.
        '500':
          description: There was an error getting the reviewee's reviews.

  /reviewees/{revieweeId}/reviews:
    post:
      summary: Create a new review for a reviewee.
      tags: 
      - Specific Reviewee
      security:
        - bearerAuth: []
      parameters: 
        - in: path
          required: true
          name: revieweeId
          description: Id of the reviewee.
          schema:
            type: string
            example: '5db492c349a67b33b8d0a2a2'
      requestBody:
        required: true
        content:
          application/json:
            schema:
             type: object
             required:
              - courseName
              - overallRating
              - recommendationRating
              - difficultyRating
              - grade
              - yearTaken
              - review
              - textbookRequired
             properties:
              courseName:
                type: string
                example: 'Brand Management 101'
              overallRating:
                type: number
                example: 4.5
              recommendationRating:
                type: number
                example: 2.5
              difficultyRating:
                type: number
                example: 3.5
              textbookRequired:
                type: boolean
                example: true
              yearTaken:
                type: integer
                example: 2015
              grade:
                type: string
                example: 'A-'
              review:
                type: string
                example: 'This class is so easy and the professor teaching style makes me understand the course material easily.'
              tags:
                type: array
                items:
                  type: string
                example: ['caring', 'young', 'easy class']
              teachingStyles:
                type: array
                items:
                  type: string
                example: ['visual', 'professional']
      responses:
        '201':
          description: Create review for a reviewee is successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '401':
          description: Authentication is required to create new review.
        '404':
          description: Reviewee not found.
        '500':
          description: There was an error creating the review.
          
  /reviewees/{revieweeId}/reviews/{reviewId}/{upOrDownVote}:
    post:
      summary: Add helpfulness vote to review.
      tags: 
      - Specific Reviewee
      security:
        - bearerAuth: []
      parameters: 
        - in: path
          required: true
          name: revieweeId
          description: Id of the reviewee being reviewed.
          schema:
            type: string
            example: '5db492c349a67b33b8d0a2a2'
        - in: path
          required: true
          name: reviewId
          description: Id of the review.
          schema:
            type: string
            example: '5db492c349a67b33b8d0a2a2'
        - in: path
          required: true
          name: upOrDownVote
          description: Option to add up or down vote.
          schema:
            type: string
            enum: ['upVote', 'downVote']
      responses:
        '201':
          description: Add helpfulness vote to review is successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '401':
          description: Authentication is required to add new vote.
        '404':
          description: Review not found.
        '500':
          description: There was an error adding helpfullness vote to the review.